requirements.txt
python-dotenv
streamlit
openai==0.27.8